+++ Server/src/ru/nemiroff/Product.java
@@ -7,11 +7,11 @@ import java.util.Map;
  */
 public class Product extends Thing {
 
-    private Map<Thing, Float> proportions;
+    private Map<Thing, Double> proportions;
 
     @Override
-    public float getPrice() {
-        float result = 0f;
+    public double getPrice() {
+        double result = 0f;
         for (Thing thing : proportions.keySet()) {
             if(proportions.get(thing) != null) {
                 result += thing.getPrice() * proportions.get(thing);
@@ -20,11 +20,11 @@ public class Product extends Thing {
         return result;
     }
 
-    public Map<Thing, Float> getProportions() {
+    public Map<Thing, Double> getProportions() {
         return proportions;
     }
 
-    public void setProportions(Map<Thing, Float> proportions) {
+    public void setProportions(Map<Thing, Double> proportions) {
         this.proportions = proportions;
     }
